requirements.txt
openai==0.28
sentence-transformers
torch
python-dotenv
